Renamed io to datasets
Moved fromimage to utils instead and load_surfer to gridder. Seemed more
appropriate.
