+++ doc/api/datasets.rst
@@ -1,0 +1,10 @@
+.. _fatiando_datasets:
+.. _fatiando_io:
+
+Fetch data from the internet (``fatiando.datasets``)
+========================================================
+
+.. automodule:: fatiando.datasets
+   :members:
+   :show-inheritance:
+
