--- doc/api/io.rst
@@ -1,9 +1,0 @@
-.. _fatiando_io:
-
-Input/Output (``fatiando.io``)
-==============================
-
-.. automodule:: fatiando.io
-   :members:
-   :show-inheritance:
-
