+++ fatiando/datasets.py
@@ -1,25 +1,15 @@
 """
-Input/Output utilities for grids, models, etc
-
-**Utilities**
-
-* :func:`~fatiando.io.fromimage`: Load array of normalized gray-scale values
-  from an image
+Load datasets from the internet.
 
 **CRUST2.0**
 
 Load and convert the `CRUST2.0 global crustal model
 <http://igppweb.ucsd.edu/~gabi/rem.html>`_ (Bassin et al., 2000).
 
-* :func:`~fatiando.io.fetch_crust2`: Download the .tar.gz archive with the model
+* :func:`~fatiando.datasets.fetch_crust2`: Download the .tar.gz archive with the model
   from the website
-* :func:`~fatiando.io.crust2_to_tesseroids`: Convert the CRUST2.0 model to
+* :func:`~fatiando.datasets.crust2_to_tesseroids`: Convert the CRUST2.0 model to
   tesseroids
-
-Input and output of regular grids in Surfer format.
-
-* :func:`~fatiando.io.load_surfer`: Read a Surfer grid file and return three 1d
-  numpy arrays and the grid shape
 
 **Sample data**
 
@@ -39,11 +29,10 @@ Surface Wave Tomography in North America, EOS Trans AGU, 81, F897, 2000.
 import urllib
 import tarfile
 
-import PIL.Image
-import scipy.misc
 import numpy
 
-import fatiando.gridder
+from . import gridder
+from .mesher import Tesseroid
 
 
 def fetch_crust2(fname='crust2.tar.gz'):
@@ -91,8 +80,6 @@ def crust2_to_tesseroids(fname):
         The converted model
 
     """
-    # Needs to be inside the function to avoid circular imports
-    from fatiando.mesher import Tesseroid
     archive = tarfile.open(fname, 'r:gz')
     # First get the topography and bathymetry information
     topogrd = _crust2_get_topo(archive)
@@ -158,115 +145,6 @@ def _crust2_get_codec(archive):
         codec[code] = {'vp':vp, 'vs':vs, 'density':density,
                        'thickness':thickness}
     return codec
-
-def fromimage(fname, ranges=None, shape=None):
-    """
-    Load an array of normalized gray-scale values from an image file.
-
-    The values will be in the range [0, 1]. The shape of the array is the shape
-    of the image (ny, nx), i.e., number of pixels in vertical (height) and
-    horizontal (width) dimensions.
-
-    Parameters:
-
-    * fname : str
-        Name of the image file
-    * ranges : [vmax, vmin] = floats
-        If not ``None``, will set the gray-scale values to this range.
-    * shape : (ny, nx)
-        If not ``None``, will interpolate the array to match this new shape
-
-    Returns:
-
-    * values : 2d-array
-        The array of gray-scale values
-
-    """
-    image = scipy.misc.fromimage(PIL.Image.open(fname), flatten=True)
-    # Invert the color scale and normalize
-    values = (image.max() - image)/numpy.abs(image).max()
-    if ranges is not None:
-        vmin, vmax = ranges
-        values *= vmax - vmin
-        values += vmin
-    if shape is not None and tuple(shape) != values.shape:
-        ny, nx = values.shape
-        X, Y = numpy.meshgrid(range(nx), range(ny))
-        values = fatiando.gridder.interp(X.ravel(), Y.ravel(), values.ravel(),
-            shape)[2].reshape(shape)
-    return values
-
-def load_surfer(fname, fmt='ascii'):
-    """
-    Read a Surfer grid file and return three 1d numpy arrays and the grid shape.
-
-    Surfer is a contouring, gridding and surface mapping software
-    from GoldenSoftware. The names and logos for Surfer and Golden
-    Software are registered trademarks of Golden Software, Inc.
-
-    http://www.goldensoftware.com/products/surfer
-
-    According to Surfer structure, x and y are horizontal and vertical
-    screen-based coordinates respectively. If the grid is in geographic
-    coordinates, x will be longitude and y latitude. If the coordinates
-    are cartesian, x will be the easting and y the norting coordinates.
-
-    WARNING: This is opposite to the convention used for Fatiando.
-    See io_surfer.py in cookbook.
-
-    Parameters:
-
-    * fname : str
-        Name of the Surfer grid file
-    * fmt : str
-        File type, can be 'ascii' or 'binary'
-
-    Returns:
-
-    * x : 1d-array
-        Value of the horizontal coordinate of each grid point.
-    * y : 1d-array
-        Value of the vertical coordinate of each grid point.
-    * grd : 1d-array
-        Values of the field in each grid point. Field can be for example
-        topography, gravity anomaly etc
-    * shape : tuple = (ny, nx)
-        The number of points in the vertical and horizontal grid dimensions,
-        respectively
-
-    """
-    assert fmt in ['ascii', 'binary'], "Invalid grid format '%s'. Should be \
-        'ascii' or 'binary'." % (fmt)
-    if fmt == 'ascii':
-        # Surfer ASCII grid structure
-        # DSAA            Surfer ASCII GRD ID
-        # nCols nRows     number of columns and rows
-        # xMin xMax       X min max
-        # yMin yMax       Y min max
-        # zMin zMax       Z min max
-        # z11 z21 z31 ... List of Z values
-        with open(fname) as ftext:
-            # DSAA is a Surfer ASCII GRD ID
-            id = ftext.readline()
-            # Read the number of columns (nx) and rows (ny)
-            nx, ny = [int(s) for s in ftext.readline().split()]
-            # Read the min/max value of x (columns/longitue)
-            xmin, xmax = [float(s) for s in ftext.readline().split()]
-            # Read the min/max value of  y(rows/latitude)
-            ymin, ymax = [float(s) for s in ftext.readline().split()]
-            # Read the min/max value of grd
-            zmin, zmax = [float(s) for s in ftext.readline().split()]
-            data = numpy.fromiter((float(i) for line in ftext for i in
-                                   line.split()), dtype='f')
-            grd = numpy.ma.masked_greater_equal(data, 1.70141e+38)
-        # Create x and y numpy arrays
-        x = numpy.linspace(xmin, xmax, nx)
-        y = numpy.linspace(ymin, ymax, ny)
-        x, y = [tmp.ravel() for tmp in numpy.meshgrid(x, y)]
-    if fmt == 'binary':
-        raise NotImplementedError("Binary file support is not implemented yet. \
-                                  Sorry")
-    return x, y, grd, (ny,nx)
 
 def fetch_bouguer_alps_egm(fname='bouguer_alps_egm08.grd'):
     """
