+++ fatiando/gridder.py
@@ -17,6 +17,11 @@ Create and operate on grids and profiles.
 * :func:`~fatiando.gridder.interp_at`
 * :func:`~fatiando.gridder.extrapolate_nans`
 
+**Input/Output**
+
+* :func:`~fatiando.gridder.load_surfer`: Read a Surfer grid file and return
+  three 1d numpy arrays and the grid shape
+
 **Misc**
 
 * :func:`~fatiando.gridder.spacing`
@@ -29,6 +34,78 @@ import numpy
 import scipy.interpolate
 import matplotlib.mlab
 
+
+def load_surfer(fname, fmt='ascii'):
+    """
+    Read a Surfer grid file and return three 1d numpy arrays and the grid shape
+
+    Surfer is a contouring, gridding and surface mapping software
+    from GoldenSoftware. The names and logos for Surfer and Golden
+    Software are registered trademarks of Golden Software, Inc.
+
+    http://www.goldensoftware.com/products/surfer
+
+    According to Surfer structure, x and y are horizontal and vertical
+    screen-based coordinates respectively. If the grid is in geographic
+    coordinates, x will be longitude and y latitude. If the coordinates
+    are cartesian, x will be the easting and y the norting coordinates.
+
+    WARNING: This is opposite to the convention used for Fatiando.
+    See io_surfer.py in cookbook.
+
+    Parameters:
+
+    * fname : str
+        Name of the Surfer grid file
+    * fmt : str
+        File type, can be 'ascii' or 'binary'
+
+    Returns:
+
+    * x : 1d-array
+        Value of the horizontal coordinate of each grid point.
+    * y : 1d-array
+        Value of the vertical coordinate of each grid point.
+    * grd : 1d-array
+        Values of the field in each grid point. Field can be for example
+        topography, gravity anomaly etc
+    * shape : tuple = (ny, nx)
+        The number of points in the vertical and horizontal grid dimensions,
+        respectively
+
+    """
+    assert fmt in ['ascii', 'binary'], "Invalid grid format '%s'. Should be \
+        'ascii' or 'binary'." % (fmt)
+    if fmt == 'ascii':
+        # Surfer ASCII grid structure
+        # DSAA            Surfer ASCII GRD ID
+        # nCols nRows     number of columns and rows
+        # xMin xMax       X min max
+        # yMin yMax       Y min max
+        # zMin zMax       Z min max
+        # z11 z21 z31 ... List of Z values
+        with open(fname) as ftext:
+            # DSAA is a Surfer ASCII GRD ID
+            id = ftext.readline()
+            # Read the number of columns (nx) and rows (ny)
+            nx, ny = [int(s) for s in ftext.readline().split()]
+            # Read the min/max value of x (columns/longitue)
+            xmin, xmax = [float(s) for s in ftext.readline().split()]
+            # Read the min/max value of  y(rows/latitude)
+            ymin, ymax = [float(s) for s in ftext.readline().split()]
+            # Read the min/max value of grd
+            zmin, zmax = [float(s) for s in ftext.readline().split()]
+            data = numpy.fromiter((float(i) for line in ftext for i in
+                                   line.split()), dtype='f')
+            grd = numpy.ma.masked_greater_equal(data, 1.70141e+38)
+        # Create x and y numpy arrays
+        x = numpy.linspace(xmin, xmax, nx)
+        y = numpy.linspace(ymin, ymax, ny)
+        x, y = [tmp.ravel() for tmp in numpy.meshgrid(x, y)]
+    if fmt == 'binary':
+        raise NotImplementedError(
+            "Binary file support is not implemented yet.")
+    return x, y, grd, (ny,nx)
 
 def regular(area, shape, z=None):
     """
